Refactor to move sql queries into db module
